Move firstOf out of flow.js
Closes: https://github.com/metarhia/metasync/issues/244
