+++ lib/flow.js
@@ -2,8 +2,6 @@
 
 const common = require('metarhia-common');
 
-const { each } = require('./array');
-
 const flow = (
   // Create a composed function from flow syntax
   fns // array of functions, callback-last / err-first
@@ -172,21 +170,8 @@ function sequential(
   next();
 }
 
-const firstOf = (
-  // Executes all asynchronous functions and pass first result to callback
-  fns, // array of function, callback-last / err-first
-  callback // function, err-first on done
-) => {
-  const done = common.once(callback);
-  each(fns, (f, iterCb) => f((...args) => {
-    done(...args);
-    iterCb(...args);
-  }));
-};
-
 module.exports = {
   flow,
   parallel,
   sequential,
-  firstOf,
 };
